move sc_timespan.hpp
Move to util/timespan.hpp
