+++ engine/buff/sc_buff.hpp
@@ -11,7 +11,7 @@
 #include <vector>
 #include <memory>
 
-#include "sc_timespan.hpp"
+#include "util/timespan.hpp"
 #include "sc_enums.hpp"
 #include "player/sc_actor_pair.hpp"
 #include "util/sample_data.hpp"
