+++ engine/item/item.hpp
@@ -10,7 +10,7 @@
 #include "dbc/item_effect.hpp"
 #include "player/gear_stats.hpp"
 #include "sc_enums.hpp"
-#include "sc_timespan.hpp"
+#include "util/timespan.hpp"
 #include "util/generic.hpp"
 #include "util/format.hpp"
 #include "util/string_view.hpp"
