+++ engine/item/special_effect.hpp
@@ -7,7 +7,7 @@
 
 #include "config.hpp"
 #include "sc_enums.hpp"
-#include "sc_timespan.hpp"
+#include "util/timespan.hpp"
 #include "util/format.hpp"
 
 #include <string>
