+++ engine/player/ground_aoe.hpp
@@ -7,7 +7,7 @@
 
 #include "config.hpp"
 #include "player_event.hpp"
-#include "sc_timespan.hpp"
+#include "util/timespan.hpp"
 #include "action/sc_action_state.hpp"
 #include <functional>
