+++ engine/sim/proc.hpp
@@ -7,7 +7,7 @@
 #include "config.hpp"
 #include "util/sample_data.hpp"
 #include "util/string_view.hpp"
-#include "sc_timespan.hpp"
+#include "util/timespan.hpp"
 
 #include <string>
