+++ engine/sim/raid_event.hpp
@@ -8,7 +8,7 @@
 #include "config.hpp"
 
 #include "sc_enums.hpp"
-#include "sc_timespan.hpp"
+#include "util/timespan.hpp"
 #include "util/generic.hpp"
 #include "util/string_view.hpp"
 #include "util/format.hpp"
