+++ engine/sim/sc_expressions.hpp
@@ -11,7 +11,7 @@
 #include <functional>
 #include <memory>
 
-#include "sc_timespan.hpp"
+#include "util/timespan.hpp"
 #include "util/span.hpp"
 #include "util/string_view.hpp"
