+++ engine/util/util.hpp
@@ -10,7 +10,7 @@
 #include "dbc/data_enums.hh"
 #include "dbc/specialization.hpp"
 #include "sc_enums.hpp"
-#include "sc_timespan.hpp"
+#include "util/timespan.hpp"
 #include "util/span.hpp"
 #include "util/string_view.hpp"
