+++ source_files/QT_engine.pri
@@ -126,7 +126,6 @@ HEADERS += engine/report/report_timer.hpp
 HEADERS += engine/report/reports.hpp
 HEADERS += engine/report/sc_highchart.hpp
 HEADERS += engine/sc_enums.hpp
-HEADERS += engine/sc_timespan.hpp
 HEADERS += engine/sim/benefit.hpp
 HEADERS += engine/sim/event.hpp
 HEADERS += engine/sim/event_manager.hpp
@@ -167,6 +166,7 @@ HEADERS += engine/util/static_map.hpp
 HEADERS += engine/util/stopwatch.hpp
 HEADERS += engine/util/string_view.hpp
 HEADERS += engine/util/timeline.hpp
+HEADERS += engine/util/timespan.hpp
 HEADERS += engine/util/util.hpp
 HEADERS += engine/util/vector_with_callback.hpp
 HEADERS += engine/util/xml.hpp
